# internal/models/token_response.go
package models

import "time"

type TokenServiceResponse struct {
	UserID       uint      `gorm:"not null"`
	Provider     string    `gorm:"not null"`
	AccessToken  string    `gorm:"not null"`
	RefreshToken string    `gorm:"null"`
	ExpirationIn int       `gorm:"not null"`
	ExpiresAt    time.Time `gorm:"not null"`
	Data         string    `gorm:"type:json;not null"`
}